Move code into private submodules, keeping existing public symbols where they are.
