+++ .travis/mypy
@@ -29,13 +29,13 @@ tmp="$(mktemp -t mypy.XXXX)";
 mypy "$@"      \
     | grep -v  \
         -e ': error: "Callable\[\[[^]]*\], [^]]*\]" has no attribute "skip"'  \
+        -e '^src/klein/_app.py:.[0-9:]*: error: All conditional function variants must have identical signatures'  \
+        -e '^src/klein/_app.py:.[0-9:]*: error: Function is missing a type annotation'  \
+        -e '^src/klein/_app.py:.[0-9:]*: error: Need type annotation for variable'  \
         -e '^src/klein/_decorators.py:.[0-9:]*: error: Function is missing a type annotation'  \
+        -e '^src/klein/_interfaces.py:.[0-9:]*: error: Function is missing a type annotation'  \
         -e '^src/klein/_plating.py:.[0-9:]*: error: Function is missing a type annotation'  \
-        -e '^src/klein/app.py:.[0-9:]*: error: All conditional function variants must have identical signatures'  \
-        -e '^src/klein/app.py:.[0-9:]*: error: Function is missing a type annotation'  \
-        -e '^src/klein/app.py:.[0-9:]*: error: Need type annotation for variable'  \
-        -e '^src/klein/interfaces.py:.[0-9:]*: error: Function is missing a type annotation'  \
-        -e '^src/klein/resource.py:.[0-9:]*: error: Function is missing a type annotation'  \
+        -e '^src/klein/_resource.py:.[0-9:]*: error: Function is missing a type annotation'  \
         -e '^src/klein/test/_trial.py:.[0-9:]*: error: Function is missing a type annotation'  \
         -e '^src/klein/test/py3_test_resource.py:.[0-9:]*: error: Function is missing a type annotation'  \
         -e '^src/klein/test/test_app.py:.[0-9:]*: error: Function is missing a type annotation'  \
