+++ src/klein/interfaces.py
@@ -1,16 +1,6 @@
-from __future__ import absolute_import, division
-
-from zope.interface import Attribute, Interface
+from ._interfaces import IKleinRequest
 
 
-class IKleinRequest(Interface):
-    branch_segments = Attribute("Segments consumed by a branch route.")
-    mapper = Attribute("L{werkzeug.routing.MapAdapter}")
-
-    def url_for(
-        self, endpoint, values=None, method=None,
-        force_external=False, append_unknown=True,
-    ):
-        """
-        L{werkzeug.routing.MapAdapter.build}
-        """
+__all__ = (
+    "IKleinRequest",
+)
